feat: Cria uma classe de erros personalizada para identificar quais são os erros gerados pela aplicação e quais são os erros lançados por ações do usuário.

+++ src/storage/group/groupCreate.ts
@@ -1,12 +1,21 @@
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import { GROUP_COLLECTION } from '@storage/storaheConfig';
-import { groupsGetAll } from './groupsGetAll';
 
-export async function groupCreate(groupName: string) {
+import { GROUP_COLLECTION } from '@storage/storaheConfig';
+
+import { groupsGetAll } from './groupsGetAll';
+import { AppError } from '@utils/AppError';
+
+export async function groupCreate(newGroup: string) {
   try {
     const storageGroups = await groupsGetAll();
 
-    const storage = JSON.stringify([...storageGroups, groupName]);
+    const groupAlreadyExists = storageGroups.includes(newGroup);
+
+    if (groupAlreadyExists) {
+      throw new AppError('Já existe um grupo com esse nome.');
+    }
+
+    const storage = JSON.stringify([...storageGroups, newGroup]);
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
   } catch (error) {
     throw error;
